+++ Repositories/src/main/java/vn/chonsoft/lixi/repositories/LixiOrderGiftRepository.java
@@ -23,7 +23,7 @@ public interface LixiOrderGiftRepository extends JpaRepository<LixiOrderGift, Lo
     @Modifying
     @Transactional
     @Query("update LixiOrderGift p set p.productId = :productId, p.productPrice = :productPrice where p.id = :id")
-    int updateProductIdAndPrice(@Param("productId") int productId, @Param("productPrice") float productPrice, @Param("id") Long id);
+    int updateProductIdAndPrice(@Param("productId") int productId, @Param("productPrice") double productPrice, @Param("id") Long id);
     
     Long deleteByOrderAndProductId(LixiOrder order, int productId);
